Moves permissions edit dialog texts to helper folder
