+++ src/app/modules/item/helpers/permissions-edit-dialog-texts.ts
@@ -1,6 +1,6 @@
 import { ProgressSelectValue } from
-  'src/app/modules/shared-components/components/collapsible-section/progress-select/progress-select.component';
-import { TypeFilter } from '../composition-filter/composition-filter.component';
+    'src/app/modules/shared-components/components/collapsible-section/progress-select/progress-select.component';
+import { TypeFilter } from '../components/composition-filter/composition-filter.component';
 
 function getTargetTypeString(targetType: TypeFilter): string {
   switch (targetType) {
